Moved some functions from statistics.js to barChart.js
